+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `index_tails` subcommand to show a list of tail numbers sorted by flight count.
 - `--geojson` argument for `add flight` to allow saving to a GeoJSON file instead of the flight log.
 
 ## [0.4.1]